nb-annotate/requirements.txt
nbformat==5.1.3
PyYAML==6.0.1
